Revert "Add capture using WinRT Windows.Graphics.Capture API. (#2149)"
This reverts commit 8f1692a5acb42eedc89ff291a6825b37a38f9ee2.

+++ src/platform/windows/input.cpp
@@ -1411,7 +1411,7 @@ namespace platf {
   ds4_update_state(gamepad_context_t &gamepad, const gamepad_state_t &gamepad_state) {
     auto &report = gamepad.report.ds4.Report;
 
-    report.wButtons = static_cast<uint16_t>(ds4_buttons(gamepad_state)) | static_cast<uint16_t>(ds4_dpad(gamepad_state));
+    report.wButtons = ds4_buttons(gamepad_state) | ds4_dpad(gamepad_state);
     report.bSpecial = ds4_special_buttons(gamepad_state);
 
     report.bTriggerL = gamepad_state.lt;
